requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0
bcrypt>=4.0
PyJWT>=2.8
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
nltk>=3.8
requests>=2.31
pypdf>=4.0
python-docx>=1.0
streamlit>=1.28
